// Frontend/ShoppingCart/src/app/models/Product/IProductModel.ts
/**
 * Defines the data model for a single product.
 */
export interface IProductModel {
  /// <summary>
  /// Uniquely identifies each product with it's unique ID.
  /// </summary>
  /// <example>1</example>
  ProductID: number;
  /// <summary>
  /// Uniquely identifies each product category with it's unique ID.
  /// </summary>
  /// <example>1</example>
  ProductCategoryID: number;
  /// <summary>
  /// Defines the category name of this product
  /// </summary>
  /// <example>Branded Foods</example> 
  ProductCategoryName: string;
  /// <summary>
  /// Defines the name of this product
  /// </summary>
  /// <example>Pepsi</example>
  ProductName: string;
  /// <summary>
  /// Defines the price of this product
  /// </summary>
  /// <example>100</example>
  ProductPrice: number;
  /// <summary>
  /// Defines the description of this product
  /// </summary>
  /// <example>A product from ABC company</example>
  ProductDescription: string;
  /// <summary>
  /// Defines the image name of this product which is autogenerated by the server
  /// </summary>
  ProductImageName: string;
  /// <summary>
  /// Defines the image API URL from the server
  /// </summary>
  /// <example>https://localhost:44398/images/1.png</example>
  ProductImageURL: string;
}